+++ workflow/scripts/monthly_time_series_2023.py
@@ -20,8 +20,6 @@ effects = pd.read_feather(snakemake.input["effects"])
 pedestrians = pd.read_feather(snakemake.input["pedestrians"])
 gps = pd.read_feather(snakemake.input["gps"])
 
-accidents["datetime"] = pd.to_datetime(accidents["p2a"], format="%d.%m.%Y")
-
 for series_name, series_metric in series_blueprint.items():
     result = series_metric.apply(
         accidents=accidents,