+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - First 2025 data
+- Metric for "Stáří motorového vozidla viníka"
+
+### Changed
+- Keys sorted in the result
 
 ## [0.0.3] - 2025-02-06